Fix AST's LibYamlWriter signature for LLP64 (Windows build)
The win/ cross build failed compiling ast/src/yamlchan.c:

  error: passing argument 2 of 'yaml_emitter_set_output' from incompatible
  pointer type [-Wincompatible-pointer-types]
    expected 'int (*)(void *, unsigned char *, size_t)'
             {aka 'int (*)(void *, unsigned char *, long long unsigned int)'}
    but argument is of type 'int (*)(void *, yaml_char_t *, long unsigned int)'

LibYamlWriter declared its size argument as `long unsigned int`, but
libyaml's yaml_write_handler_t uses `size_t`. On LP64 - macOS, Linux -
those are the same 64-bit type, so it compiles everywhere we had tested.
mingw is LLP64: `long` is 32 bits and `size_t` is 64, so the function
pointer genuinely does not match.

Two things kept this hidden. The types coincide on every platform this had
been built on, and it only became fatal with GCC 14 (theirs is 14.4.0),
which promoted -Wincompatible-pointer-types from a warning to an error -
the `-w` the ast recipe already passes suppresses warnings, not errors.

Fixed by using size_t, which is also what the sibling LibYamlReader has
always used - the writer was simply inconsistent with it. This is upstream
AST code, not ours, and it breaks any Windows build of AST with YAML
enabled, so it is noted in TODO.md as something to send upstream alongside
the version-ceiling patches.

Verified on macOS, where the change is semantically a no-op: libast
recompiles and relinks clean, and the GWCS read path still returns the
ground-truth corner exactly (269.986945543501 65.974265112478).

+++ ast/src/yamlchan.c
@@ -505,7 +505,7 @@ static int Get1I( AstKeyMap *, const char *, int, int, int *, int *, int * );
 static int GetChoice( AstKeyMap *, const char *, const char *, int, int, int * );
 static int IsA( AstKeyMap *, const char *, int * );
 static int LibYamlReader( void *, yaml_char_t *, size_t, size_t * );
-static int LibYamlWriter( void *, yaml_char_t *, long unsigned int );
+static int LibYamlWriter( void *, yaml_char_t *, size_t );
 static int ReadBaseFrame( AstKeyMap *, AstSkyFrame *, int * );
 static int SimplifyAsdf( AstYamlChan *, AstKeyMap **, int *);
 static int Use( AstYamlChan *, int, int, int * );
@@ -6355,7 +6355,7 @@ static int LibYamlReader( void *data, yaml_char_t *buffer, size_t size,
 }
 
 static int LibYamlWriter( void *data, yaml_char_t *buffer,
-                          long unsigned int size ){
+                          size_t size ){
 /*
 *  Name:
 *     LibYamlWriter
@@ -6369,7 +6369,7 @@ static int LibYamlWriter( void *data, yaml_char_t *buffer,
 *  Synopsis:
 *     #include "yamlchan.h"
 *     int LibYamlWriter( void *data, yaml_char_t *buffer,
-*                        long unsigned int size )
+*                        size_t size )
 
 *  Class Membership:
 *     YamlChan member function.
